Day 16

+++ Source/handmade.cpp
@@ -44,8 +44,8 @@ RenderWeirdGradient(game_offscreen_buffer *Buffer, int BlueOffset, int GreenOffs
             // ACTUAL Pixel in Memory: BB GG RR XX
             // So all windows bitmaps have Blue First, Green then Red Bytes!
 
-            uint8 Blue  = (X + BlueOffset);
-            uint8 Green = (Y + GreenOffset);
+            uint8 Blue  = (uint8)(X + BlueOffset);
+            uint8 Green = (uint8)(Y + GreenOffset);
 
             *Pixel++ = ((Green << 8) | Blue);
         }
@@ -83,8 +83,8 @@ GameUpdateAndRender(game_memory *Memory,
     game_controller_input *Input0 = &Input->Controllers[0];
     if(Input0->IsAnalog)
     {
-        GameState->BlueOffset += (int)4.0f * (Input0->EndX);
-        GameState->ToneHz = 256 + (int)(128.0f * (Input0->EndY));
+        GameState->BlueOffset += (int)(4.0f * Input0->EndX);
+        GameState->ToneHz = 256 + (int)(128.0f * Input0->EndY);
     }
     else
     {
